dev-core-services/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1py
dj-rest-auth==5.0.2 #5.0.1
Django==4.2
django-allauth==0.59.0 #0.55.2
django-cors-headers==4.3.1 #4.3.0
django-otp==1.3.0 
django-qr-code==4.0.1
django-qrcode==0.3
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1 #5.3.0
idna==3.6
oauthlib==3.2.2
pillow==10.2.0 #9.5.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==6.1 #7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
segno==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0


#psycopg2-binary==2.9.9,